+++ modules/blockchain/src/reference.rs
@@ -37,8 +37,7 @@ pub enum Error {
     #[error("The block could not apply successfully")]
     Ledger {
         #[source]
-        #[from]
-        source: ledger::Error,
+        source: Box<ledger::Error>,
     },
 
     #[error("The block's epoch validity failed to successfully apply")]
@@ -82,7 +81,11 @@ impl Reference {
     /// another blockchain or may have a specific meaning
     pub fn new(block0: &Block) -> Result<Self, Error> {
         let header = block0.header().clone();
-        let ledger = Ledger::new(header.hash(), block0.contents().iter_slice())?;
+        let ledger = Ledger::new(header.hash(), block0.contents().iter_slice()).map_err(|e| {
+            Error::Ledger {
+                source: Box::new(e),
+            }
+        })?;
         let epoch_info = Arc::new(EpochInfo::new(block0, &ledger)?);
         let previous_epoch_state = None;
 
@@ -180,14 +183,19 @@ impl Reference {
 
         transition_state.epoch_info.check_header(block.header())?;
 
-        let ledger = transition_state.ledger().apply_block(
-            transition_state
-                .epoch_info
-                .epoch_ledger_parameters()
-                .clone(),
-            block.contents(),
-            &metadata,
-        )?;
+        let ledger = transition_state
+            .ledger()
+            .apply_block(
+                transition_state
+                    .epoch_info
+                    .epoch_ledger_parameters()
+                    .clone(),
+                block.contents(),
+                &metadata,
+            )
+            .map_err(|e| Error::Ledger {
+                source: Box::new(e),
+            })?;
 
         Ok(Self {
             ledger,
@@ -202,18 +210,27 @@ impl Reference {
     /// and distribute the rewards
     pub fn epoch_transition(&self) -> Result<Self, Error> {
         // 1. apply protocol changes
-        let ledger = self.ledger.apply_protocol_changes()?;
+        let ledger = self
+            .ledger
+            .apply_protocol_changes()
+            .map_err(|e| Error::Ledger {
+                source: Box::new(e),
+            })?;
         // 2. distribute rewards
         let ledger = if let Some(distribution) = self
             .epoch_info
             .epoch_leadership_schedule()
             .stake_distribution()
         {
-            let (ledger, _rewards) = ledger.distribute_rewards(
-                distribution,
-                self.epoch_info.epoch_ledger_parameters(),
-                RewardsInfoParameters::default(),
-            )?;
+            let (ledger, _rewards) = ledger
+                .distribute_rewards(
+                    distribution,
+                    self.epoch_info.epoch_ledger_parameters(),
+                    RewardsInfoParameters::default(),
+                )
+                .map_err(|e| Error::Ledger {
+                    source: Box::new(e),
+                })?;
             ledger
         } else {
             ledger