+++ testing/hersir/src/controller/monitor/node/mod.rs
@@ -222,8 +222,7 @@ impl Node {
         self.process
             .wait_for_shutdown(Duration::from_secs(150))
             .map_err(|e| {
-                self.progress_bar
-                    .log_info(format!("shutdown error: {}", e.to_string()));
+                self.progress_bar.log_info(format!("shutdown error: {}", e));
                 Error::NodeFailedToShutdown {
                     alias: self.alias(),
                     e,
@@ -233,16 +232,6 @@ impl Node {
                 self.progress_bar.log_info("shutdown successfully.");
                 exit_status
             })
-    }
-
-    fn ports_are_opened(&self) -> bool {
-        self.port_opened(self.process.rest_address().port())
-            && self.port_opened(self.process.p2p_listen_addr().port())
-    }
-
-    fn port_opened(&self, port: u16) -> bool {
-        use std::net::TcpListener;
-        TcpListener::bind(("127.0.0.1", port)).is_ok()
     }
 
     pub fn is_up(&self) -> bool {