+++ testing/jormungandr-integration-tests/src/non_functional/network/mod.rs
@@ -1,3 +1,4 @@
+pub mod big;
 pub mod desync;
 pub mod disruption;
 #[cfg(feature = "soak")]